// src/app/core/services/theme/interfaces/app-theme.model.ts
import { ComponentThemeConfig } from './component-theme.model';

export interface AppTheme {
  body: ComponentThemeConfig;
  header: ComponentThemeConfig;
  headerButton: ComponentThemeConfig;
  button: ComponentThemeConfig;
  card: ComponentThemeConfig;
  componentSelector: ComponentThemeConfig;
  closeButton: ComponentThemeConfig;
  editor: ComponentThemeConfig;
  input: ComponentThemeConfig;
  label: ComponentThemeConfig;
  skillCard: ComponentThemeConfig;
}